Implement automated eclipse build
A customized version of the pde-maven-plugin is used to generate
the feature.xml based on Maven dependencies.

The implementation uses Tycho to generate the executable.
Ultimately this should be a simple task, but Tycho requires everything
to use its packaging e.g. eclipse-plugin, eclipse-feature, eclipse-repository.

Maven profiles are used to isolate Tycho from the rest of the build
and to build the same artifacts using regular + Tycho packaging.

The project can be built with:
mvn clean install && mvn install -Ptycho